Corregido

+++ RANK1/FT_PRINTF/ft_printf.h
@@ -6,15 +6,15 @@
 /*   By: rmarin-n <rmarin-n@student.42barcelona.co  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/29 11:22:52 by rmarin-n          #+#    #+#             */
-/*   Updated: 2025/10/06 15:07:45 by rmarin-n         ###   ########.fr       */
+/*   Updated: 2025/10/09 15:33:03 by rmarin-n         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
-#include <unistd.h>
-#include <stdarg.h>
+# include <unistd.h>
+# include <stdarg.h>
 
 /* +=====================+
  * |     ft_printf.c     |
@@ -36,7 +36,7 @@ int		ft_printf_u(unsigned int u);
  * | ft_printf_utils2.c  |
  * +=====================+ */
 
-int		ft_printf_x(unsigned int x, int up);
-int		ft_printf_p(void* p);
+int		ft_printf_x(unsigned long x, int up);
+int		ft_printf_p(void *p);
 
 #endif
